# use LoopVectorization and custom single pass newton's method to speed up `irr` 2x to 5x