--- a/Project.toml
+++ b/Project.toml
@@ -4,11 +4,12 @@ authors = ["alecloudenback <alecloudenback@users.noreply.github.com> and contrib
 version = "1.0.1"
 
 [deps]
+LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 
 [compat]
-julia = "1.6"
 Roots = "^1.0, 2"
+julia = "1.6"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"