--- a/Project.toml
+++ b/Project.toml
@@ -8,6 +8,7 @@ LoopVectorization = "bdcacae8-1622-11e9-2a5c-532679323890"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
 
 [compat]
+LoopVectorization = "^0.12"
 Roots = "^1.0, 2"
 julia = "1.6"