--- a/src/FinanceCore.jl
+++ b/src/FinanceCore.jl
@@ -1,5 +1,6 @@
 module FinanceCore
 import Roots
+using LoopVectorization
 
 include("AbtractYield.jl")