```julia
# original
julia> @benchmark irr($cfs,0:50)
BenchmarkTools.Trial: 10000 samples with 7 evaluations.
 Range (min … max):  4.494 μs …  12.649 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.643 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.757 μs ± 420.592 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃  ▃█▁▂▂▁                                                   ▁
  █▆▂██████▆▆▆▇█▆▆▆▆█▇▆▇▇▇▇▆▇▇▇▆▆▆▆▆▇▇▆▅▅▆▅▆▅▆▅▅▅▄▅▄▅▄▅▅▅▆▆▆▆ █
  4.49 μs      Histogram: log(frequency) by time      6.48 μs <

 Memory estimate: 32 bytes, allocs estimate: 1.

julia> @benchmark irr($v[1:3])
BenchmarkTools.Trial: 10000 samples with 192 evaluations.
 Range (min … max):  509.115 ns …  14.000 μs  ┊ GC (min … max): 0.00% … 95.37%
 Time  (median):     529.297 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   545.506 ns ± 261.249 ns  ┊ GC (mean ± σ):  0.94% ±  1.90%

  ▂▄▄▃ ▁▇█▆▄▄▁▁▃▃▂▁▁▁  ▁    ▂▂▂▁  ▁▁▁                           ▂
  ████▇██████████████████▇▇██████▇████▇▇▇▇▇▇▇▇▇█▆▅▆▆▆▅▅▆▅▆▅▅▅▅▅ █
  509 ns        Histogram: log(frequency) by time        670 ns <

 Memory estimate: 112 bytes, allocs estimate: 2.

# new 

julia> @benchmark irr($cfs,0:50)
BenchmarkTools.Trial: 10000 samples with 134 evaluations.
 Range (min … max):  722.634 ns …  16.890 μs  ┊ GC (min … max): 0.00% … 94.91%
 Time  (median):     744.403 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   751.382 ns ± 166.466 ns  ┊ GC (mean ± σ):  0.21% ±  0.95%

  █▃▄▂  █▅▅▃▁  ▃ ▁                                              ▂
  ████████████▇███▇██▇▇▇▇▇▆▇▇▆▆▅▇▇▇▇█▇▇▇▇▆▆▆▆▆▆▆▇▇▆▆▅▅▅▅▅▅▄▄▄▄▆ █
  723 ns        Histogram: log(frequency) by time        935 ns <

 Memory estimate: 32 bytes, allocs estimate: 1.

julia> @benchmark irr($v[1:3])
BenchmarkTools.Trial: 10000 samples with 570 evaluations.
 Range (min … max):  204.751 ns …   5.397 μs  ┊ GC (min … max): 0.00% … 95.17%
 Time  (median):     206.505 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   219.286 ns ± 181.659 ns  ┊ GC (mean ± σ):  3.07% ±  3.56%

  ▇█▄▁▁▂▅▆▂   ▂▃                           ▁                    ▁
  ████████████████▇▇▇████▇▇██████▇█████▇█████▇▇▇▆▇▆▆▇▆▆▆▅▅▆▅▅▄▆ █
  205 ns        Histogram: log(frequency) by time        263 ns <

 Memory estimate: 112 bytes, allocs estimate: 2.
```